code cleaning

+++ srcs/set_dead_time.c
@@ -6,7 +6,7 @@
 /*   By: tsekiguc <tsekiguc@student.42tokyo.jp      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/16 22:49:11 by tsekiguc          #+#    #+#             */
-/*   Updated: 2021/12/20 23:15:13 by tsekiguc         ###   ########.fr       */
+/*   Updated: 2021/12/27 10:39:29 by tsekiguc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,15 +15,14 @@
 void	set_dead_time(t_philo *philo)
 {
 	struct timeval	tv;
-	int				tmp;
-	int				sec;
-	int				msec;
+	long			tmp;
+	long			sec;
+	long			msec;
 
 	gettimeofday(&tv, NULL);
 	tmp = (tv.tv_usec / 1000) + philo->args->time_to_die;
 	sec = tmp / 1000;
 	msec = tmp % 1000;
-
 	pthread_mutex_lock(philo->lock);
 	philo->dead.sec = tv.tv_sec + sec;
 	philo->dead.msec = msec;
